Feat: Smart Schedule for events
I added the entire Smart Scheduler system to create a similar arrangement to reclaim (I will improve it further in the use-calendar.tsx file)
First, the events in the tuplan calendar will be similar to the events in reclaim AI. Reclaim AI does not classify based on work hour, personal hour or meeting hour (this is the task that James is working on) but based on event properities such as family, personal, education and will have a way that it relies on the user's settings for event properities time (optimal time and preferred days)
First, I added smart scheduling settings to create event properities time for users to adjust for activities (note that overlapping events are not resolved yet).
This system will arrange based on first the event properities for the task then the tasks with high priority go into the optimal hours according to the user's settings), the medium will replace the empty slots on the calendar and finally the low. (need to adjust again after completing other tasks) --> similar to reclaim's arrangement